Renames route to relatorios.
